Added serializable_id to many classes, hashCode override, DTO class

+++ src/main/java/ru/ssau/tk/DoubleA/javalabs/persistence/entity/Calculation.java
@@ -2,8 +2,6 @@ package ru.ssau.tk.DoubleA.javalabs.persistence.entity;
 
 import jakarta.persistence.*;
 
-import java.util.Set;
-
 @Entity
 @Table (name = "calculation")
 public class Calculation {
@@ -11,7 +9,7 @@ public class Calculation {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column (name = "calculation_id")
-    private long id;
+    private int id;
 
     @Column (name = "applied_value", nullable = false)
     private double appliedX;
@@ -22,11 +20,11 @@ public class Calculation {
     @Column (name = "calculation_hash", nullable = false)
     private long hash;
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
